Refactor paginated class
